Add files via upload
He conseguido insertar el apellido en el listado de personas, pero el id de la categoria no consigo visualizarlo. Y la eliminación de personas no se hace al hacer click en el boton de eliminar de la tabla del listado de personas.

+++ persona-ficha.php
@@ -9,20 +9,33 @@ $nuevaPersona = ($id == -1);
 
 if ($nuevaPersona) {
     $nombrePersona = "<introduzca nombre>";
+    $apellidosPersona = "<introduzca apellidos>";
     $telefonoPersona = "<introduzca su teléfono>";
     $idCategoria = "<introduzca el id de la categoría a la que pertenece>";
 } else {
-    $sql = "SELECT nombre, telefono, categoria_id FROM persona WHERE id=?";
+    $sqlPersonas = "SELECT nombre, apellidos, telefono, categoriaId FROM persona WHERE id=?";
 
-    $select = $conexion->prepare($sql);
-    $select->execute([]);
-    $resultSet = $select->fetchAll();
+    $select = $conexion->prepare($sqlPersonas);
+    $select->execute([$id]);
+    $resultSetPersona = $select->fetchAll();
 
-    $nombrePersona = $resultSet[0]["nombre"];
-    $telefonoPersona = $resultSet[1]["telefono"];
-    $idCategoria = $resultSet[2]["categoria_id"];
+    $nombrePersona = $resultSetPersona[0]["nombre"];
+    $apellidosPersona = $resultSetPersona[0]["apellidos"];
+    $telefonoPersona = $resultSetPersona[0]["telefono"];
+    $idCategoria = $resultSetPersona[0]["categoriaId"];
 }
+
+// Dejamos preparado un recordset con las categorías.
+
+$sqlCategorias = "SELECT id, nombre FROM categoria ORDER BY nombre";
+
+$select = $conexion->prepare($sqlCategorias);
+$select->execute([]); // Array vacío porque la consulta preparada no requiere parámetros.
+$resultSetCategorias = $select->fetchAll();
+
+
 ?>
+
 
 
 <html>
@@ -42,13 +55,15 @@ if ($nuevaPersona) {
 <?php } ?>
 
 <form method='post' action='persona-guardar.php'>
-
     <input type='hidden' name='id' value='<?=$id?>' />
-
     <ul>
         <li>
             <strong>Nombre: </strong>
             <input type='text' name='nombre' value='<?=$nombrePersona?>' />
+        </li>
+        <li>
+            <strong> Apellidos: </strong>
+            <input type='text' name='apellidos' value='<?=$apellidosPersona?>' />
         </li>
         <li>
             <strong>Teléfono: </strong>
@@ -56,7 +71,7 @@ if ($nuevaPersona) {
         </li>
         <li>
             <strong>ID-Categoría: </strong>
-            <input type='number' name='categoria_id' value='<?=$idCategoria?>'>
+            <input type='number' name='categoriaId' value='<?=$idCategoria?>'>
         </li>
     </ul>
 
@@ -65,16 +80,14 @@ if ($nuevaPersona) {
     <?php } else { ?>
         <input type='submit' name='guardar' value='Guardar cambios' />
     <?php } ?>
-
 </form>
-
 <br>
 
-<a href="persona-eliminar.php?id=<?=$id ?>">Eliminar persona</a>
+<?php if (!$nuevaPersona){ ?>
+    <a href="persona-eliminar.php?id=<?=$id ?>">Eliminar persona</a>
+<?php } ?>
 
 <br>
-<br>
-
 <a href="persona-listado.php">Volver a la lista de Personas.</a>
 
 </body>
